refactor: move all settings components to their own file
