+++ lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:dynamic_color/dynamic_color.dart';
-import 'package:flutter/foundation.dart';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -72,6 +71,7 @@ class _SpritpreiseState extends State<Spritpreise> {
         ),
         darkTheme: ThemeData(
           colorScheme: darkColorScheme ?? Spritpreise._defaultDarkColorScheme,
+          appBarTheme: AppBarTheme(backgroundColor: Spritpreise._defaultDarkColorScheme.background),
           scaffoldBackgroundColor: Spritpreise._defaultDarkColorScheme.background,
           useMaterial3: true,
         ),
