+++ lib/src/components/settings/settings_group.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+class SettingsGroup extends StatelessWidget {
+  const SettingsGroup({super.key, required this.title, required this.children});
+
+  final String title;
+  final List<Widget> children;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: List.from([
+        Row(
+          children: [
+            Padding(
+              padding: const EdgeInsets.only(left: 16, top: 16),
+              child: Text(
+                title,
+                style: TextStyle(
+                    color: Theme.of(context).colorScheme.primary,
+                    fontWeight: FontWeight.bold,
+                    fontSize: DefaultTextStyle.of(context).style.fontSize! * 17 / 20),
+              ),
+            ),
+          ],
+        )
+      ])
+        ..addAll(children),
+    );
+  }
+}
